tests,refactoring and ts errors
