+++ packages/frontend/cypress.json
@@ -1,4 +1,6 @@
 {
   "baseUrl": "http://localhost:3000",
-  "supportFile": "cypress/support/index.ts"
+  "supportFile": "cypress/support/index.ts",
+  "viewportWidth": 1600,
+  "viewportHeight": 900
 }
