+++ packages/frontend/src/components/AddUsers_Modal/AddUsers_Modal.test.tsx
@@ -2,7 +2,7 @@ import { Provider } from "react-redux";
 import configureStore from "redux-mock-store";
 import AddUsers_Modal from "../AddUsers_Modal";
 import renderer from "react-test-renderer";
-import { AuthState } from "../../redux/reducer/authReducer";
+import { AuthState } from "../../redux/reducer/authReducer/authReducer";
 import { render, screen } from "@testing-library/react";
 import "@testing-library/jest-dom";
 import React from "react";
