+++ packages/frontend/src/components/Avatar/Group_avatar/Group_avatar.test.tsx
@@ -1,4 +1,5 @@
 import Group_avatar from "../Group_avatar";
+import React from "react";
 import renderer from "react-test-renderer";
 import { render, screen } from "@testing-library/react";
 import "@testing-library/jest-dom";
@@ -7,6 +8,15 @@ const setupRender = () => {
   const component = render(<Group_avatar inviter={"ivan"} cookieName={"greg"} members={["ivan", "greg"]} />);
   return component;
 };
+
+jest.mock("react", () => {
+  const originReact = jest.requireActual("react");
+  const mUseRef = jest.fn();
+  return {
+    ...originReact,
+    useRef: mUseRef,
+  };
+});
 
 describe("Render connected React-redux page", () => {
   beforeEach(() => {
