+++ packages/frontend/src/components/ChatRoom/ChatRoom.test.tsx
@@ -4,8 +4,8 @@ import { render, screen } from "@testing-library/react";
 import "@testing-library/jest-dom";
 import configureStore from "redux-mock-store";
 import { Provider } from "react-redux";
-import { initialState as setState } from "../../redux/reducer/setReducer";
-import { initialState as messageState } from "../../redux/reducer/messageReducer";
+import { initialState as setState } from "../../redux/reducer/setReducer/setReducer";
+import { initialState as messageState } from "../../redux/reducer/messageReducer/messageReducer";
 import Cookies from "universal-cookie";
 
 const mockStore = configureStore([]);
@@ -38,6 +38,7 @@ describe("Render connected React-redux page", () => {
   });
 
   it("should render <ChatRoom/>", () => {
+    screen.getByRole("");
     render(
       <Provider store={store}>
         <ChatRoom cookie={cookie} chatId={"321312313"} />
