+++ packages/frontend/src/components/MainSection/MainSection.test.tsx
@@ -3,8 +3,8 @@ import MainSection from "./MainSection";
 import renderer from "react-test-renderer";
 import { render, screen } from "@testing-library/react";
 import { Provider } from "react-redux";
-import { initialState } from "../../redux/reducer/setReducer";
-import { initialState as saveInputState } from "../../redux/reducer/save_inputReducer";
+import { initialState } from "../../redux/reducer/setReducer/setReducer";
+import { initialState as saveInputState } from "../../redux/reducer/save_inputReducer/save_inputReducer";
 import MockedSocket from "socket.io-mock";
 import configureStore from "redux-mock-store";
 import Cookies from "universal-cookie";
