+++ packages/frontend/src/components/RegisterForm/QuickLogin_Modal/QuickLogin_Modal.tsx
@@ -8,7 +8,7 @@ import { useDispatch } from "react-redux";
 const QuickLogin_Modal = ({ quickLogin, isLogging }: { quickLogin: () => void; isLogging: boolean }) => {
   const dispatch = useDispatch();
   return (
-    <div style={{ width: "75%", height: "30vh" }} className="fRequests_modal flex">
+    <div style={{ width: "75%", height: "30vh", position: "absolute", zIndex: "200" }} className="fRequests_modal flex">
       <div
         className={cx(
           css`
