+++ packages/frontend/src/components/RegisterForm/RegisterForm.test.tsx
@@ -3,8 +3,8 @@ import renderer from "react-test-renderer";
 import { render, screen } from "@testing-library/react";
 import "@testing-library/jest-dom";
 import { Provider } from "react-redux";
-import { initialState } from "../../redux/reducer/authReducer";
-import { initialState as saveInputState } from "../../redux/reducer/save_inputReducer";
+import { initialState } from "../../redux/reducer/authReducer/authReducer";
+import { initialState as saveInputState } from "../../redux/reducer/save_inputReducer/save_inputReducer";
 import configureStore from "redux-mock-store";
 const quickLogin: any = jest.fn();
 const handleSubmit = jest.fn();
