+++ packages/frontend/src/components/UserSettings/UserSettings.test.tsx
@@ -2,7 +2,7 @@ import "@testing-library/jest-dom";
 import { render, screen } from "@testing-library/react";
 import UserSettings from "./UserSettings";
 import { Provider } from "react-redux";
-import { initialState } from "../../redux/reducer/setReducer";
+import { initialState } from "../../redux/reducer/setReducer/setReducer";
 import configureStore from "redux-mock-store";
 import renderer from "react-test-renderer";
 const mockStore = configureStore([]);
