+++ packages/frontend/src/redux/reducer/authReducer/state.ts
@@ -1,0 +1,7 @@
+export interface IAuthState {
+  remember_me: boolean;
+  loginPrompt: boolean;
+  good?: string;
+  bad?: string;
+  cookie?: string;
+}
