+++ packages/frontend/src/redux/reducer/index.ts
@@ -1,9 +1,9 @@
 import { combineReducers } from "redux";
-import reducer from "./authReducer";
-import setReducer from "./setReducer";
-import saveInputReducer from "./save_inputReducer";
-import messageReducer from "./messageReducer";
-import { ActionType } from "../state";
+import reducer from "./authReducer/authReducer";
+import setReducer from "./setReducer/setReducer";
+import saveInputReducer from "./save_inputReducer/save_inputReducer";
+import messageReducer from "./messageReducer/messageReducer";
+import { ActionType } from "../types";
 const combiReducers = combineReducers({
   authReducer: reducer,
   setReducer,
