+++ packages/frontend/src/redux/reducer/messageReducer/state.ts
@@ -1,0 +1,6 @@
+import { IchatInstance } from "../../../components/ChatRoom/ChatRoom";
+
+export interface InitialStateMessage {
+  messages: IchatInstance[];
+  show: boolean;
+}
