+++ packages/frontend/src/redux/reducer/save_inputReducer/save_inputReducer.ts
@@ -1,13 +1,16 @@
-import { Action, ActionType, InitialState3 } from "../state";
+import ISave_inputState from "./state";
 import { HYDRATE } from "next-redux-wrapper";
-export const initialState = {
+import { AnyAction } from "redux";
+import { ActionType } from "../../types";
+
+export const initialState: ISave_inputState | ISave_inputState[] = {
   input_username: "",
   input_password: "",
   input_email: "",
   input_gender: "",
   notification_number: 0,
 };
-const saveInputReducer = (state: InitialState3 = initialState, action: Action) => {
+const saveInputReducer = (state = initialState, action: AnyAction) => {
   switch (action.type) {
     case HYDRATE:
       return {
