+++ packages/frontend/src/redux/reducer/save_inputReducer/state.ts
@@ -1,0 +1,7 @@
+export default interface ISave_inputState {
+  input_username: string;
+  input_password: string;
+  input_email: string;
+  input_gender: string;
+  notification_number: number;
+}
