+++ packages/frontend/src/redux/reducer/setReducer/state.ts
@@ -1,0 +1,13 @@
+export interface IInitialSet {
+  reciever: string;
+  chat_inviter: string;
+  pageNumber: number;
+  setUserSettings: boolean;
+  setFriendRequest: boolean;
+  toggleCreateGroup: boolean;
+  setMobileNav: boolean;
+  setIsMatch: boolean;
+  setChatSettings: boolean;
+  setModalInvite: boolean;
+  setIsLoggingIn: boolean;
+}
