+++ packages/frontend/src/redux/store.ts
@@ -2,9 +2,9 @@ import { createStore, applyMiddleware } from "redux";
 import { createWrapper } from "next-redux-wrapper";
 import { composeWithDevTools } from "redux-devtools-extension";
 import thunk from "redux-thunk";
-import reducers from "./reducer";
+import combReducers from "./reducer";
 
-const makeStore = () => createStore(reducers, {}, composeWithDevTools(applyMiddleware(thunk)));
+const makeStore = () => createStore(combReducers, {}, composeWithDevTools(applyMiddleware(thunk)));
 
 export const wrapper = createWrapper(makeStore, { debug: true });
 export * as actions from "./actions/authActions";
